Changed classs hierarchy
Task is now a different class contained within the Leaf class (which is
of type node, part of the tree)

Every tasks inherits from the Task class. The trees are made of nodes
(Composite and Leaf) and the Leaf contain a task (one task instance per
tree, different leaves with same task for same tree).

After a few tests, combination seem to be working.